### 2025-07-01 b806bcc
Updating pydantic with DOI
