+++ paper/paper.bib
@@ -1,11 +1,17 @@
+### https://zenodo.org/records/15662245
+### Colvin, S., Jolibois, E., Ramezani, H., Garcia Badaracco, A., Dorsey, T., Montague, D.,
+###   Matveenko, S., Trylesinski, M., Runkle, S., Hewitt, D., Hall, A., & Plot, V. (2025). 
+###   Pydantic (v2.11.7). Zenodo. https://doi.org/10.5281/zenodo.15662245
 
-@misc{Colvin:2017,
-  author = {Samuel Colvin},
-  title = {Pydantic},
-  year = {2017},
-  howpublished = {\url{https://github.com/pydantic/pydantic}},
-  note = {GitHub repository},
+@software{pydantic,
+  author       = {Colvin, Samuel and Jolibois, Eric and Ramezani, Hasan and Garcia Badaracco, Adrian and Dorsey, Terrence and Montague, David and Matveenko, Serge and Trylesinski, Marcelo and Runkle, Sydney and Hewitt, David and Hall, Alex and Plot, Victorien},
+  title        = {Pydantic (v2.11.7)},
+  year         = {2025},
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.15662245},
+  url          = {https://doi.org/10.5281/zenodo.15662245}
 }
+
 @misc{Gronholm:2016,
   author = {Alex Grönholm},
   title = {Typeguard},
