+++ paper/paper.bib
@@ -12,41 +12,43 @@
   url          = {https://doi.org/10.5281/zenodo.15662245}
 }
 
-@misc{Gronholm:2016,
-  author = {Alex Grönholm},
-  title = {Typeguard},
-  year = {2016},
-  howpublished = {\url{https://github.com/agronholm/typeguard}},
-  note = {GitHub repository},
+#### Updating to make these software entries
+@software{Gronholm:2016,
+  author    = {Alex Gr{\"o}nholm},
+  title     = {Typeguard},
+  version   = {4.4.4},
+  year      = {2016},
+  url       = {https://github.com/agronholm/typeguard},
+  urldate   = {2025-07-01},
+  note      = {Run-time type checker for Python; GitHub repository}
 }
-@misc{Lehtosalo:2012,
-  author = {Jukka Lehtosalo},
-  title = {Mypy},
-  year = {2012},
-  howpublished = {\url{https://github.com/python/mypy}},
-  note = {GitHub repository},
+
+@software{Lehtosalo:2012,
+  author    = {Jukka Lehtosalo and the Mypy contributors},
+  title     = {Mypy},
+  version   = {1.16.1},
+  year      = {2012},
+  url       = {https://github.com/python/mypy},
+  urldate   = {2025-07-01},
+  note      = {Static type checker for Python; GitHub repository}
 }
-@misc{KeithMagee:2016,
-  author = {Russell Keith-Magee},
-  title = {Enforce},
-  year = {2016},
-  howpublished = {\url{https://github.com/Russell91/python-enforce}},
-  note = {GitHub repository},
+
+@software{KeithMagee:2016,
+  author    = {Russell Keith‑Magee},
+  title     = {Enforce},
+  version   = {0.3.4},
+  year      = {2016},
+  url       = {https://github.com/RussBaz/enforce},
+  urldate   = {2025-07-01},
+  note      = {Runtime type hint enforcement for Python; GitHub repository}
 }
-@misc{Loria:2013,
-  author = {Steven Loria},
-  title = {Marshmallow},
-  year = {2013},
-  howpublished = {\url{https://github.com/marshmallow-code}},
-  note = {GitHub repository},
-}
-@inproceedings{Rak-amnouykit:2020,
-  author = {Ingkarat Rak-amnouykit and Daniel McCrevan and Ana Milanova and Martin Hirzel and Julian Dolby},
-  title = {Python 3 Types in the Wild: A Tale of Two Type Systems},
-  booktitle = {Proceedings of the 16th ACM SIGPLAN International Symposium on Dynamic Languages (DLS 2020)},
-  year = {2020},
-  pages = {59--72},
-  publisher = {ACM},
-  doi = {10.1145/3426422.3426984},
-  url = {https://hirzels.com/martin/papers/dls20-python-types.pdf},
+
+@software{Loria:2013,
+  author    = {Steven Loria},
+  title     = {Marshmallow},
+  version   = {4.0.0},
+  year      = {2013},
+  url       = {https://github.com/marshmallow-code/marshmallow},
+  urldate   = {2025-07-01},
+  note      = {Object serialization library for Python; GitHub repository}
 }
