### 2025-07-01 6ed4ca3
Changed reference for pydantic
