+++ paper/paper.bib
@@ -3,12 +3,13 @@
 ###   Matveenko, S., Trylesinski, M., Runkle, S., Hewitt, D., Hall, A., & Plot, V. (2025). 
 ###   Pydantic (v2.11.7). Zenodo. https://doi.org/10.5281/zenodo.15662245
 
-@software{pydantic,
+@software{Colvin:2017,
   author       = {Colvin, Samuel and Jolibois, Eric and Ramezani, Hasan and Garcia Badaracco, Adrian and Dorsey, Terrence and Montague, David and Matveenko, Serge and Trylesinski, Marcelo and Runkle, Sydney and Hewitt, David and Hall, Alex and Plot, Victorien},
   title        = {Pydantic (v2.11.7)},
-  year         = {2025},
+  year         = {2017},
   publisher    = {Zenodo},
   doi          = {10.5281/zenodo.15662245},
+  urldate      = {2025-07-01},
   url          = {https://doi.org/10.5281/zenodo.15662245}
 }
 
