### 2025-07-01 159be04
Update bib (trussell merge).
