+++ paper/paper.bib
@@ -1,46 +1,55 @@
+### https://zenodo.org/records/15662245
+### Colvin, S., Jolibois, E., Ramezani, H., Garcia Badaracco, A., Dorsey, T., Montague, D.,
+###   Matveenko, S., Trylesinski, M., Runkle, S., Hewitt, D., Hall, A., & Plot, V. (2025). 
+###   Pydantic (v2.11.7). Zenodo. https://doi.org/10.5281/zenodo.15662245
 
-@misc{Colvin:2017,
-  author = {Samuel Colvin},
-  title = {Pydantic},
-  year = {2017},
-  howpublished = {\url{https://github.com/pydantic/pydantic}},
-  note = {GitHub repository},
+@software{Colvin:2017,
+  author       = {Colvin, Samuel and Jolibois, Eric and Ramezani, Hasan and Garcia Badaracco, Adrian and Dorsey, Terrence and Montague, David and Matveenko, Serge and Trylesinski, Marcelo and Runkle, Sydney and Hewitt, David and Hall, Alex and Plot, Victorien},
+  title        = {Pydantic (v2.11.7)},
+  year         = {2017},
+  publisher    = {Zenodo},
+  doi          = {10.5281/zenodo.15662245},
+  urldate      = {2025-07-01},
+  url          = {https://doi.org/10.5281/zenodo.15662245}
 }
-@misc{Gronholm:2016,
-  author = {Alex Grönholm},
-  title = {Typeguard},
-  year = {2016},
-  howpublished = {\url{https://github.com/agronholm/typeguard}},
-  note = {GitHub repository},
+
+#### Updating to make these software entries
+@software{Gronholm:2016,
+  author    = {Alex Gr{\"o}nholm},
+  title     = {Typeguard},
+  version   = {4.4.4},
+  year      = {2016},
+  url       = {https://github.com/agronholm/typeguard},
+  urldate   = {2025-07-01},
+  note      = {Run-time type checker for Python; GitHub repository}
 }
-@misc{Lehtosalo:2012,
-  author = {Jukka Lehtosalo},
-  title = {Mypy},
-  year = {2012},
-  howpublished = {\url{https://github.com/python/mypy}},
-  note = {GitHub repository},
+
+@software{Lehtosalo:2012,
+  author    = {Jukka Lehtosalo and the Mypy contributors},
+  title     = {Mypy},
+  version   = {1.16.1},
+  year      = {2012},
+  url       = {https://github.com/python/mypy},
+  urldate   = {2025-07-01},
+  note      = {Static type checker for Python; GitHub repository}
 }
-@misc{KeithMagee:2016,
-  author = {Russell Keith-Magee},
-  title = {Enforce},
-  year = {2016},
-  howpublished = {\url{https://github.com/Russell91/python-enforce}},
-  note = {GitHub repository},
+
+@software{KeithMagee:2016,
+  author    = {Russell Keith‑Magee},
+  title     = {Enforce},
+  version   = {0.3.4},
+  year      = {2016},
+  url       = {https://github.com/RussBaz/enforce},
+  urldate   = {2025-07-01},
+  note      = {Runtime type hint enforcement for Python; GitHub repository}
 }
-@misc{Loria:2013,
-  author = {Steven Loria},
-  title = {Marshmallow},
-  year = {2013},
-  howpublished = {\url{https://github.com/marshmallow-code}},
-  note = {GitHub repository},
-}
-@inproceedings{Rak-amnouykit:2020,
-  author = {Ingkarat Rak-amnouykit and Daniel McCrevan and Ana Milanova and Martin Hirzel and Julian Dolby},
-  title = {Python 3 Types in the Wild: A Tale of Two Type Systems},
[…]
