### 2025-07-03 cef8314
Allow for strict to be disabled in type enforcement: #58
