### 2025-07-03 d10bc68
Add benchmarking and update readme.md
