+++ test/benchmark.py
@@ -1,0 +1,162 @@
+try:
+    import time, sys
+    from typing import Union, Dict, List
+    from statistics import mean
+
+    from beartype import beartype
+    from typeguard import typechecked
+    import type_enforced
+
+    from pydantic import BaseModel, validate_call
+
+    # Open the log file, clear it and redirect stdout to it
+    log = open("benchmark.md", "w")
+    sys.stdout.flush()  # Ensure the log file is cleared before writing
+    sys.stdout = log
+
+    REPEATS = 100
+
+    # --- Test data
+    five_key_dict = {f'key{i}': i for i in range(5)}
+    big_key_dict = {f'key{i}': i for i in range(1000)}
+
+    five_item_list = [1, 2.0, 3, 4.0, 5]
+    big_item_list = [float(i) if i % 2 else i for i in range(1000)]
+
+    # --- Benchmark and Validation test cases
+    test_cases = {
+        "int": (42, "not an int"),
+        "Union[int,float]": (3.14, "not a number"),
+        "str": ("hello", 123),
+        "dict[str,int] (5 keys)": (five_key_dict, {"k1": 1, "k2": "two", "k3": 3}),
+        "dict[str,int] (1000 keys)": (big_key_dict, {"k1": 1, "k2": "two", "k3": 3}),
+        "list[Union[int,float]] (5 items)": (five_item_list, [1, "two", 3, 4, 5]),
+        "list[Union[int,float]] (1000 items)": (big_item_list, [1, "two", 3, 4, 5]*200),
+        "list[dict[str,int]] (5 items)": ([five_key_dict] * 5, [{"k1": 1, "k2": "two", "k3": 3}]),
+        "list[dict[str,int]] (100 items)": ([big_key_dict] * 100, [{"k1": 1, "k2": "two", "k3": 3}]),
+    }
+
+    # --- Typing definitions
+    types = {
+        "int": int,
+        "Union[int,float]": Union[int, float],
+        "str": str,
+        "dict[str,int] (5 keys)": Dict[str, int],
+        "dict[str,int] (1000 keys)": Dict[str, int],
+        "list[Union[int,float]] (5 items)": List[Union[int, float]],
+        "list[Union[int,float]] (1000 items)": List[Union[int, float]],
+        "list[dict[str,int]] (5 items)": List[Dict[str, int]],
+        "list[dict[str,int]] (100 items)": List[Dict[str, int]],
+    }
+
+    # --- Timing helper
+    def timeit(func, arg):
+        durations = []
+        for _ in range(REPEATS):
+            start = time.perf_counter()
+            func(arg)
+            durations.append(time.perf_counter() - start)
+        return mean(durations) * 1e6  # microseconds
+
+    # --- Factory functions
+    def pydantic_factory(typ):
+        class PModel(BaseModel):
+            x: typ
+
+        @validate_call
+        def f(x: typ) -> None:
+            PModel(x=x)
+        return f
+
+    def beartype_factory(typ):
+        @beartype
+        def f(x: typ) -> None:
+            pass
+        return f
+
+    def typeguard_factory(typ):
+        @typechecked
+        def f(x: typ) -> None:
+            pass
+        return f
+
+    def type_enforced_factory(typ):
+        @type_enforced.Enforcer()
+        def f(x: typ) -> None:
+            pass
+        return f
+
+    # --- Checkers and factories
+    checkers = {
+        "type_enforced": type_enforced_factory,
+        "Pydantic": pydantic_factory,
+        "Beartype": beartype_factory,
+        "Typeguard": typeguard_factory,
+    }
+
+    # --- Validation helper
+    def test_validation(func, valid_value, invalid_value):
+        try:
+            func(valid_value)
+            valid_passed = True
+        except Exception:
[…]
