+++ requirements.txt
@@ -4,4 +4,7 @@ build>=1.2.2
 pdoc>=15.0.1
 twine>=6.1.0
 virtualenv>=20.29.2
+pydantic
+beartype
+typeguard
 -e .

[…]
