+++ test/benchmark.py
@@ -17,8 +17,8 @@ try:
     REPEATS = 100
 
     # --- Test data
-    five_key_dict = {f'key{i}': i for i in range(5)}
-    big_key_dict = {f'key{i}': i for i in range(1000)}
+    five_key_dict = {f"key{i}": i for i in range(5)}
+    big_key_dict = {f"key{i}": i for i in range(1000)}
 
     five_item_list = [1, 2.0, 3, 4.0, 5]
     big_item_list = [float(i) if i % 2 else i for i in range(1000)]
@@ -28,12 +28,30 @@ try:
         "int": (42, "not an int"),
         "Union[int,float]": (3.14, "not a number"),
         "str": ("hello", 123),
-        "dict[str,int] (5 keys)": (five_key_dict, {"k1": 1, "k2": "two", "k3": 3}),
-        "dict[str,int] (1000 keys)": (big_key_dict, {"k1": 1, "k2": "two", "k3": 3}),
-        "list[Union[int,float]] (5 items)": (five_item_list, [1, "two", 3, 4, 5]),
-        "list[Union[int,float]] (1000 items)": (big_item_list, [1, "two", 3, 4, 5]*200),
-        "list[dict[str,int]] (5 items)": ([five_key_dict] * 5, [{"k1": 1, "k2": "two", "k3": 3}]),
-        "list[dict[str,int]] (100 items)": ([big_key_dict] * 100, [{"k1": 1, "k2": "two", "k3": 3}]),
+        "dict[str,int] (5 keys)": (
+            five_key_dict,
+            {"k1": 1, "k2": "two", "k3": 3},
+        ),
+        "dict[str,int] (1000 keys)": (
+            big_key_dict,
+            {"k1": 1, "k2": "two", "k3": 3},
+        ),
+        "list[Union[int,float]] (5 items)": (
+            five_item_list,
+            [1, "two", 3, 4, 5],
+        ),
+        "list[Union[int,float]] (1000 items)": (
+            big_item_list,
+            [1, "two", 3, 4, 5] * 200,
+        ),
+        "list[dict[str,int]] (5 items)": (
+            [five_key_dict] * 5,
+            [{"k1": 1, "k2": "two", "k3": 3}],
+        ),
+        "list[dict[str,int]] (100 items)": (
+            [big_key_dict] * 100,
+            [{"k1": 1, "k2": "two", "k3": 3}],
+        ),
     }
 
     # --- Typing definitions
@@ -66,24 +84,28 @@ try:
         @validate_call
         def f(x: typ) -> None:
             PModel(x=x)
+
         return f
 
     def beartype_factory(typ):
         @beartype
         def f(x: typ) -> None:
             pass
+
         return f
 
     def typeguard_factory(typ):
         @typechecked
         def f(x: typ) -> None:
             pass
+
         return f
 
     def type_enforced_factory(typ):
         @type_enforced.Enforcer()
         def f(x: typ) -> None:
             pass
+
         return f
 
     # --- Checkers and factories
@@ -112,27 +134,44 @@ try:
 
     # --- Final output
     print(f"# Benchmark Results (python {sys.version.split(' ')[0]})\n")
-    print("This file contains the results of the benchmark tests for various type checkers.\n")
+    print(
+        "This file contains the results of the benchmark tests for various type checkers.\n"
+    )
     print("Generated by /test/benchmark.py\n")
     print("Each checker is tested with different data types and structures")
     print("- Every checker gets the same data and test cases")
-    print(f"- The average time taken to validate the data over {REPEATS} runs is measured.")
+    print(
+        f"- The average time taken to validate the data over {REPEATS} runs is measured."
+    )
     print("\n## Results Summary")
-    print("The following table summarizes the average time taken by each type checker for different data types and structures.\n")
[…]
