+++ type_enforced/__init__.py
@@ -26,15 +26,21 @@ pip install type_enforced
 ```py
 import type_enforced
 
-@type_enforced.Enforcer(enabled=True)
+@type_enforced.Enforcer(enabled=True, strict=True)
 def my_fn(a: int , b: int | str =2, c: int =3) -> None:
     pass
 ```
-- Note: `enabled=True` by default if not specified. You can set `enabled=False` to disable type checking for a specific function, method, or class. This is useful for a production vs debugging environment or for undecorating a single method in a larger wrapped class.
+- Note: `enabled=True` by default if not specified. You can set `enabled=False` to disable type checking for a specific function, method, or class. This is useful for a production vs debugging environment or for undecorating a single method in a larger wrapped class. 
+- Note: `strict=True` by default if not specified. You can set `strict=False` to disable exceptions being raised when type checking fails. Instead, a warning will be printed to the console.
 
 ## Getting Started
 
 `type_enforcer` contains a basic `Enforcer` wrapper that can be used to enforce many basic python typing hints. [Technical Docs Here](https://connor-makowski.github.io/type_enforced/type_enforced/enforcer.html).
+
+`Enforcer` can be used as a decorator for functions, methods, and classes. It will enforce the type hints on the function or method inputs and outputs. It takes in the following optional arguments:
+
+- `enabled` (True): A boolean to enable or disable type checking. If `True`, type checking will be enforced. If `False`, type checking will be disabled.
+- `strict` (True): A boolean to enable or disable type mismatch exceptions. If `True` exceptions will be raised when type checking fails. If `False`, exceptions will not be raised but instead a warning will be printed to the console.
 
 `type_enforcer` currently supports many single and multi level python types. This includes class instances and classes themselves. For example, you can force an input to be an `int`, a number `int | float`, an instance of the self defined `MyClass`, or a even a vector with `list[int]`. Items like `typing.List`, `typing.Dict`, `typing.Union` and `typing.Optional` are supported.
 
@@ -44,27 +50,33 @@ Nesting is allowed as long as the nested items are iterables (e.g. `typing.List`
 
 Variables without an annotation for type are not enforced.
 
-## What changed in 2.0.0?
-The main changes in version 2.0.0 revolve around migrating towards the standard python typing hint process and away from the original type_enfoced type hints (as type enforced was originally created before the `|` operator was added to python).
-- Support for python3.10 has been dropped.
-- List based union types are no longer supported.
-    - For example `[int, float]` is no longer a supported type hint.
-    - Use `int|float` or `typing.Union[int, float]` instead.
-- Dict types now require two types to be specified.
-    - The first type is the key type and the second type is the value type.
-    - For example, `dict[str, int|float]` or `dict[int, float]` are valid types.
-- Tuple types now allow for `N` types to be specified.
-    - Each item refers to the positional type of each item in the tuple.
-    - Support for ellipsis (`...`) is supported if you only specify two types and the second is the ellipsis type.
-        - For example, `tuple[int, ...]` or `tuple[int|str, ...]` are valid types.
-    - Note: Unions between two tuples are not supported.
-        - For example, `tuple[int, str] | tuple[str, int]` will not work.
-- Constraints and Literals can now be stacked with unions.
-    - For example, `int | Constraint(ge=0) | Constraint(le=5)` will require any passed values to be integers that are greater than or equal to `0` and less than or equal to `5`.
-    - For example, `Literal['a', 'b'] | Literal[1, 2]` will require any passed values that are equal (`==`) to `'a'`, `'b'`, `1` or `2`.
-- Literals now evaluate during the same time as type checking and operate as OR checks.
[…]
