### 2025-07-07 f85969e
Update readme for why use type enforced.
