### 2025-08-01 #8517 @nkrusch
@dorchard would you be interested in reviewing this Python type enforcer for JOSS?
