### 2025-08-05 676d4c0
Update readme for downloads to use pepy.tech as data source instead of pypistats.org
