### 2025-08-15 3c75ce4
Fix partial for better wrapper use.
