+++ type_enforced/utils.py
@@ -32,7 +32,11 @@ class Partial:
         new_args = self.__args__ + args
         new_kwargs = {**self.__kwargs__, **kwargs}
         # Create a comprehensive set of assigned variable names to determine arity
-        assigned_vars = set(self.__fn_arg_default_keys__ + self.__fn_arg_keys__[:len(new_args)] + list(new_kwargs.keys()))
+        assigned_vars = set(
+            self.__fn_arg_default_keys__
+            + self.__fn_arg_keys__[: len(new_args)]
+            + list(new_kwargs.keys())
+        )
         arity = self.__fnArity__ - len(assigned_vars)
         if arity < 0:
             self.__exception__("Too many arguments were supplied")
@@ -50,13 +54,19 @@ class Partial:
         Get the default values of the passed function or method and store them in `self.__fn_defaults__`.
         """
         self.__fn_var_keys__ = list(self.__fn__.__code__.co_varnames)
-        self.__fn_arg_keys__ = self.__fn_var_keys__[: self.__fn__.__code__.co_argcount]
+        self.__fn_arg_keys__ = self.__fn_var_keys__[
+            : self.__fn__.__code__.co_argcount
+        ]
         if self.__fn__.__defaults__ is not None:
-            self.__fn_arg_default_keys__ = self.__fn_arg_keys__[-len(self.__fn__.__defaults__):]
+            self.__fn_arg_default_keys__ = self.__fn_arg_keys__[
+                -len(self.__fn__.__defaults__) :
+            ]
         else:
             self.__fn_arg_default_keys__ = []
         if self.__fn__.__kwdefaults__ is not None:
-            self.__fn_arg_default_keys__.extend(list(self.__fn__.__kwdefaults__.keys()))
+            self.__fn_arg_default_keys__.extend(
+                list(self.__fn__.__kwdefaults__.keys())
+            )
 
     def __get__(self, instance, owner):
         def bind(*args, **kwargs):
