+++ setup.cfg
@@ -1,6 +1,6 @@
 [metadata]
 name = type_enfoced
-version = 2.2.0
+version = 2.2.1
 description_file = README.md
 
 [options]
