### 2025-08-19 #8832 @editorialbot
[REVIEW]: Type Enforced: A Python type enforcer for type annotations
**Submitting author:** @connor-makowski (<a href="http://orcid.org/0009-0005-1522-022X">Connor Makowski</a>)
**Repository:** https://github.com/connor-makowski/type_enforced
**Branch with paper.md** (empty if default branch): main
**Version:** 2.2.3
