### 2025-08-26 897bb6d
Backtrack to pypistats.org download badge.
