### 2025-08-28 a148efd
Add support for future refs #82
