### 2025-08-28 fd8d754
Change test_class_15 to avoid future import and instead use quoted future type ref.
