+++ test/test_class_15.py
@@ -1,11 +1,9 @@
-from __future__ import annotations
-
 import type_enforced
 
 
 class Foo:
     @type_enforced.Enforcer
-    def __init__(self, object: Bar) -> None:
+    def __init__(self, object: "Bar") -> None:
         pass
 
 
