### 2025-09-01 connor-makowski/type_enforced#63 @amcandio
JOSS review: authorship clarification
I noticed that a few co-authors do not appear to have commits in the repository. I understand that JOSS allows non-code contributions to justify authorship, such as project direction, design, or documentation. For transparency and to ensure alignment with JOSS authorship criteria, could you briefly clarify the contributions of each co-author?

https://github.com/openjournals/joss-reviews/issues/8832
