### 2025-09-01 connor-makowski/type_enforced#64 @amcandio
JOSS review: README improvements
A couple comments on the readme:
- README should clarify "runtime" on the first sentence: "A pure python (no special compiler required) type enforcer for type annotations. Enforce types in python functions and methods."
- There is an odd empty code-block after "Note: See the example below or technical".
- Cosmetical (you can ignore): consider using aliases when importing, for example `import type_enforced as te`

https://github.com/openjournals/joss-reviews/issues/8832
