### 2025-09-01 connor-makowski/type_enforced#65 @amcandio
JOSS review: constrains typing and static typing
Using constraints as part of the typing system `(int | Constraint(ge=2)` is an interesting idea but it can have some undesirable side-effects on static typing (especially if the return type has constraints declared). This is because typing system might not be able to tell the exact type being returned (i.e. is it an int or is it a constraint?).

This should be mentioned in the docs/paper. I don't think there is an easy way to solve it. I thought about using dynamically generated types but  mypy doesn't handle those quite well and I couldn't make it work.

https://github.com/openjournals/joss-reviews/issues/8832
