### 2025-09-01 connor-makowski/type_enforced#66 @amcandio
JOSS review: Community guidelines
For point "Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support":

### 1. Contribute to the software

Here it would be great to have more details on how to run the package locally. It mentions installing docker, so pointers to installation would be great. However, since this is a pure-python library, using virtualenv would provide a simpler contributor experience. Any chance we can drop docker dependency?

###  2. Report issues or problems with the software 3. Seek support

It would be great to add mention to those in the README.

https://github.com/openjournals/joss-reviews/issues/8832
