### 2025-09-01 connor-makowski/type_enforced#67 @amcandio
JOSS review: manuscript review
Here is a list of comments on the manustcript:
- When looking at latest pdf for the paper I see a bunch of "(?)" when pointing to references.
- Typo: unitialized -> uninitialized?
- There is mention to "minimal overhead" but no data is provided to back that claim. It would be great to add some experiments on this so users can be aware of the overhead in advance. I'm particularly curious about nexted types, which might require multiple validation checks.
- Constraints handlying should have its own section. Notation such as `int | Constraint(ge=0)` can catch readers by surprise. A comment to https://github.com/connor-makowski/type_enforced/issues/65 here would be great.
- What exactly are "inheritance-aware validation" and "uninitialized class type checks?
- Research Applications could be condensed and merged into Statemenf of need. In case you prefer it as its own section, it could be strengthened by including specific, concrete examples of use cases or scenarios.

https://github.com/openjournals/joss-reviews/issues/8832
