### 2025-09-01 #8832 @amcandio
I just finished my first pass and opened some issues. This is a great library.
