### 2025-09-03 connor-makowski/type_enforced#64 @connor-makowski
@amcandio Good points.
- Runtime to readme: Makes perfect sense. I will fix this.
- Empty code block: Also makes sense. I will fix this too.
- Cosmetic: As you implied, this one is preferential. I have gone back and forth on this one and could not make a decision so I ended up leaning towards more explicit. For consistency, I think I will skip this one for now, but will keep in mind longer term as I work on more nascent projects.

Side note: I am getting pretty far behind main branch with the JOSS branch. Should I make changes in JOSS? Can I merge from main into JOSS and then make changes? That, however really changes a lot of things in the repo.

Suggestions are very welcome as this is our first JOSS paper.
