### 2025-09-03 connor-makowski/type_enforced#66 @connor-makowski
@amcandio Good catch. Will get those added.
