### 2025-09-03 connor-makowski/type_enforced#64 @amcandio
I prefer to use the main because it's the default and sometimes I forget to switch. Do let me know if I should be looking at Joss instead!
