### 2025-09-05 connor-makowski/type_enforced#67 @connor-makowski
@mitrussell Can you help me out here?
