### 2025-09-05 f6ed289
Update constraint docs to better differentiate it from standard typing for #65.
