### 2025-09-05 0be7470
Add Support and Development sections to the main readme. #66
