### 2025-09-05 connor-makowski/type_enforced#66 @connor-makowski
@amcandio Seems like this should be on track with the above commit.
