### 2025-09-07 connor-makowski/type_enforced#66 @amcandio
Sweet, closing this one
