### 2025-09-09 3b4a57b
Updates related to Issue #67
