### 2025-09-10 ac8ea3d
Fix for #66: Confirmed working with docker and venv on linux.
