+++ utils/test.sh
@@ -1,6 +1,7 @@
 #!/bin/bash
 python --version
-for file in /app/test/*.py; do
+
+for file in "$(dirname "$0")"/../test/*.py; do
     [ -e "$file" ] || continue  # Skip if no files match
     python "$file"
 done
