### 2025-09-10 fbc5a7f
Readme cleanup for #64.
