### 2025-09-16 connor-makowski/type_enforced#67 @connor-makowski
@amcandio We are working on this.
- ?. Not sure how this slipped through. Working on this.
- Typo: Fixed with: #69
- Minimal overhead: Good point. We should probably clean up that language to be more explicit to what we intended.
- Nexted types: I am not sure if these are even supported. 😱 I can double check this.
- Constraints: Handled in #65
- "inheritance-aware validation" and "uninitialized class type checks": These are clear in the main README with examples, but not specified well in the paper. Good catch. Will add some additional verbiage.
- Research applications: Good catch. Merged into statement of need in #65

@mitrussell @mrkwanzaa outstanding work for us:
- `?` in references
- Minimal Overhead Cleanup
- Nexted Types: Check and possibly benchmark
-  "inheritance-aware validation" and "uninitialized class type checks"
