### 2025-09-20 c67d894
Adding responses to Issue #67
