+++ paper/paper.bib
@@ -1,7 +1,7 @@
 ### https://zenodo.org/records/15662245
 ### Colvin, S., Jolibois, E., Ramezani, H., Garcia Badaracco, A., Dorsey, T., Montague, D.,
 ###   Matveenko, S., Trylesinski, M., Runkle, S., Hewitt, D., Hall, A., & Plot, V. (2025). 
-###   Pydantic (v2.11.7). Zenodo. https://doi.org/10.5281/zenodo.15662245
+###   Pydantic (v2.11.7). Zenodo. https://doi.org/10.5281/zenodo.15662245 
 
 @software{Colvin:2017,
   author       = {Colvin, Samuel and Jolibois, Eric and Ramezani, Hasan and Garcia Badaracco, Adrian and Dorsey, Terrence and Montague, David and Matveenko, Serge and Trylesinski, Marcelo and Runkle, Sydney and Hewitt, David and Hall, Alex and Plot, Victorien},
@@ -52,4 +52,23 @@
   url       = {https://github.com/marshmallow-code/marshmallow},
   urldate   = {2025-07-01},
   note      = {Object serialization library for Python; GitHub repository}
+}
+
+@inproceedings{Rak-amnouykit:2020,
+	location = {Virtual {USA}},
+	title = {Python 3 types in the wild: a tale of two type systems},
+	isbn = {978-1-4503-8175-8},
+	url = {https://dl.acm.org/doi/10.1145/3426422.3426981},
+	doi = {10.1145/3426422.3426981},
+	shorttitle = {Python 3 types in the wild},
+	abstract = {Python 3 is a highly dynamic language, but it has introduced a syntax for expressing types with {PEP}484. This paper explores how developers use these type annotations, the type system semantics provided by type checking and inference tools, and the performance of these tools. We evaluate the types and tools on a corpus of public {GitHub} repositories. We review {MyPy} and {PyType}, two canonical static type checking and inference tools, and their distinct approaches to type analysis. We then address three research questions: (i) How often and in what ways do developers use Python 3 types? (ii) Which type errors do developers make? (iii) How do type errors from different tools compare? Surprisingly, when developers use static types, the code rarely type-checks with either of the tools. {MyPy} and {PyType} exhibit false positives, due to their static nature, but also flag many useful errors in our corpus. Lastly, {MyPy} and {PyType} embody two distinct type systems, flagging different errors in many cases. Understanding the usage of Python types can help guide tool-builders and researchers. Understanding the performance of popular tools can help increase the adoption of static types and tools by practitioners, ultimately leading to more correct and more robust Python code.},
+	eventtitle = {{SPLASH} '20: Conference on Systems, Programming, Languages, and Applications, Software for Humanity},
+	pages = {57--70},
+	booktitle = {Proceedings of the 16th {ACM} {SIGPLAN} International Symposium on Dynamic Languages},
+	publisher = {{ACM}},
+	author = {Rak-amnouykit, Ingkarat and {McCrevan}, Daniel and Milanova, Ana and Hirzel, Martin and Dolby, Julian},
+	urldate = {2025-09-20},
+  year      = {2020},
+	date = {2020-11-17},
+	langid = {english},
 }
