### 2025-09-20 connor-makowski/type_enforced#67 @mitrussell
@amcandio Thank you for your review. Please find below a summary of our responses and the corresponding revisions:

References displaying as "(?)":
All instances of "(?)" placeholders have been replaced with appropriate citations from our reference list. In particular, the work of Rak-amnouykit et al. (2020) has been cited to support statements regarding static type checking and type inference tools in Python. It had been inadvertently left from our paper.bib file.

Typo ("unitialized" → "uninitialized"):
This typo has been corrected.

Minimal overhead claim:
We revised this claim and pointed to the benchmarks in this repo.

Constraints handling section:
Constraints handling is now presented in its own dedicated section of the README, with updated notation and references. We specifically address recent discussion in the JOSS review system and clarify the usage of constructs such as int | Constraint(ge=0).

"Inheritance-aware validation" and "uninitialized class type checks":
These features are now described in the "Key Features" section.

Research Applications and Statement of Need:
The "Research Applications" section has been condensed and merged into the "Statement of Need."

We appreciate your feedback and believe these revisions have strengthened the clarity of our manuscript. Thank you for helping us improve the paper.
