### 2025-09-20 connor-makowski/type_enforced#67 @amcandio
I did another pass and the paper looks really good. I'm closing this one.
