+++ paper/paper.bib
@@ -46,7 +46,7 @@
 
 @software{Loria:2013,
   author    = {Steven Loria},
-  title     = {Marshmallow},
+  title     = {{marshmallow}},
   version   = {4.0.0},
   year      = {2013},
   url       = {https://github.com/marshmallow-code/marshmallow},
@@ -56,7 +56,7 @@
 
 @inproceedings{Rak-amnouykit:2020,
 	location = {Virtual {USA}},
-	title = {Python 3 types in the wild: a tale of two type systems},
+	title = {{Python 3 Types in the Wild: A Tale of Two Type Systems}},
 	isbn = {978-1-4503-8175-8},
 	url = {https://dl.acm.org/doi/10.1145/3426422.3426981},
 	doi = {10.1145/3426422.3426981},
@@ -71,4 +71,24 @@
   year      = {2020},
 	date = {2020-11-17},
 	langid = {english},
+}
+
+@software{pyright:2025,
+  author    = {Microsoft Corporation},
+  title     = {Pyright},
+  version   = {1.1.405},
+  year      = {2025},
+  url       = {https://github.com/microsoft/pyright},
+  urldate   = {2025-09-21},
+  note      = {Static Type Checker for Python; GitHub repository}
+}
+
+@software{pytype:2025,
+  author    = {Google},
+  title     = {Pytype},
+  version   = {2024.10.11},
+  year      = {2024},
+  url       = {https://github.com/google/pytype},
+  urldate   = {2025-09-21},
+  note      = {A static type analyzer for Python code; GitHub repository}
 }
