### 2025-09-21 8014892
bib format edit
Signed-off-by: Neea <neearusch@gmail.com>
