+++ paper/paper.bib
@@ -74,21 +74,21 @@
 }
 
 @software{pyright:2025,
-  author    = {Microsoft Corporation},
-  title     = {Pyright},
-  version   = {1.1.405},
-  year      = {2025},
-  url       = {https://github.com/microsoft/pyright},
-  urldate   = {2025-09-21},
-  note      = {Static Type Checker for Python; GitHub repository}
+	title        = {Pyright},
+	author       = {{Microsoft Corporation}},
+	year         = 2025,
+	url          = {https://github.com/microsoft/pyright},
+	urldate      = {2025-09-21},
+	note         = {Static Type Checker for Python; GitHub repository},
+	version      = {1.1.405}
 }
 
 @software{pytype:2025,
-  author    = {Google},
-  title     = {Pytype},
-  version   = {2024.10.11},
-  year      = {2024},
-  url       = {https://github.com/google/pytype},
-  urldate   = {2025-09-21},
-  note      = {A static type analyzer for Python code; GitHub repository}
+	title        = {Pytype},
+	author       = {Google},
+	year         = 2024,
+	url          = {https://github.com/google/pytype},
+	urldate      = {2025-09-21},
+	note         = {A static type analyzer for Python code; GitHub repository},
+	version      = {2024.10.11}
 }
