### 2025-09-22 170f118
Lock external type check repo deps for benchmark testing.
