### 2025-09-23 888e896
2.2.3: Bump version
