+++ type_enforced/__init__.py
@@ -5,7 +5,7 @@
 [![PyPI Downloads](https://img.shields.io/pypi/dm/type_enforced.svg?label=PyPI%20downloads)](https://pypi.org/project/type_enforced/)
 <!-- [![PyPI Downloads](https://pepy.tech/badge/type_enforced)](https://pypi.org/project/type_enforced/) -->
 
-A pure python (no special compiler required) type enforcer for type annotations. Enforce types in python functions and methods.
+A pure python runtime type enforcer for type annotations. Enforce types in python functions and methods.
 
 # Setup
 
@@ -31,7 +31,7 @@ import type_enforced
 def my_fn(a: int , b: int | str =2, c: int =3) -> None:
     pass
 ```
-- Note: `enabled=True` by default if not specified. You can set `enabled=False` to disable type checking for a specific function, method, or class. This is useful for a production vs debugging environment or for undecorating a single method in a larger wrapped class. 
+- Note: `enabled=True` by default if not specified. You can set `enabled=False` to disable type checking for a specific function, method, or class. This is useful for a production vs debugging environment or for undecorating a single method in a larger wrapped class.
 - Note: `strict=True` by default if not specified. You can set `strict=False` to disable exceptions being raised when type checking fails. Instead, a warning will be printed to the console.
 
 ## Getting Started
@@ -53,13 +53,13 @@ Variables without an annotation for type are not enforced.
 
 ## Why use Type Enforced?
 
-- `type_enforced` is a pure python type enforcer that does not require any special compiler or preprocessor to work. 
+- `type_enforced` is a pure python type enforcer that does not require any special compiler or preprocessor to work.
 - `type_enforced` uses the standard python typing hints and enforces them at runtime.
     - This means that you can use it in any python environment (3.11+) without any special setup.
 - `type_enforced` is designed to be lightweight and easy to use, making it a great choice for both small and large projects.
 - `type_enforced` supports complex (nested) typing hints, union types, and many of the standard python typing functions.
-- `type_enforced` is designed to be fast and efficient, with minimal overhead. 
-- `type_enforced` offers the fastest performance for enforcing large objects of complex types 
+- `type_enforced` is designed to be fast and efficient, with minimal overhead.
+- `type_enforced` offers the fastest performance for enforcing large objects of complex types
     - Note: See the [benchmarks](https://github.com/connor-makowski/type_enforced/blob/main/benchmark.md) for more information on the performance of each type checker.
 
 ## Supported Type Checking Features:
@@ -124,6 +124,7 @@ Variables without an annotation for type are not enforced.
 - `Constraint` validation.
     - This is a special type of validation that allows passed input to be validated.
         - Standard and custom constraints are supported.
+    - Constraints are not actually types. They are type_enforced specific validators and may cause issues with other runtime or static type checkers like `mypy`.
     - This is useful for validating that a passed input is within a certain range or meets a certain criteria.
     - Note: Constraints stack when used with unions.
         - e.g. `int | Constraint(ge=0) | Constraint(le=5)` will require any passed values to be integers that are greater than or equal to `0` and less than or equal to `5`.
@@ -131,7 +132,6 @@ Variables without an annotation for type are not enforced.
         - This operates differently than other checks (like `Literal`) and is evaluated post type checking.
         - For example, if you have an annotation of `str | Constraint(ge=0)`, this will always raise an exception since if you pass a string, it will raise on the constraint check and if you pass an integer, it will raise on the type check.
[…]
