### 2025-09-25 #8832 @connor-makowski
Thanks for these recommendations @crvernon.

The Marshmallow line is fixed: fb1adb690d4cc76f94af7b9eac5948a6bac9d372

Follow up questions for how the PDF is generated. Right now, the paper MD file goes:

```
Recent empirical studies, such as [@Rak-amnouykit:2020], have analyzed the adoption and semantics of Python's type systems in real-world codebases, highlighting both the promise and the challenges of practical type enforcement.
```

I am not sure of a way to remove the parentheses since they are inserted automatically. I could do something like:

```
Recent empirical studies, such as one by Rak-amnouykit [@Rak-amnouykit:2020], have analyzed the adoption and semantics of Python's type systems in real-world codebases, highlighting both the promise and the challenges of practical type enforcement.
```
