### 2025-09-25 #8832 @crvernon
Hi @connor-makowski , you can fix this by dropping the brackets...so in your example:

```
Recent empirical studies, such as @Rak-amnouykit:2020, have analyzed the adoption and semantics of Python's type systems in real-world codebases, highlighting both the promise and the challenges of practical type enforcement.
```
