### 2025-09-25 b7dec2a
Remove citation () in paper.
