### 2025-09-25 #8832 @connor-makowski
Thanks. Super helpful! I appreciate the quick response very much.

See: https://github.com/connor-makowski/type_enforced/commit/b7dec2a5755cde8f7c7fc11206a34bff0dcc754d
