### 2025-09-26 #8832 @connor-makowski
@crvernon 🤦 I totally missed that one. Thanks for the follow up.

Fixed in: https://github.com/connor-makowski/type_enforced/commit/b3eabc35090d9ce90751ba27f2710fef723dabc3
